@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
-import { FormControl, Validators } from '@angular/forms';
+import { UntypedFormControl, Validators } from '@angular/forms';
 import { EmailServiceService } from 'src/app/services/email-service.service';
 
 @Component({
@@ -9,9 +9,9 @@ import { EmailServiceService } from 'src/app/services/email-service.service';
   styleUrls: ['./contact-page.component.css']
 })
 export class ContactPageComponent implements OnInit {
-  name = new FormControl('',Validators.required);
-  email = new FormControl('',[Validators.required, Validators.email]);
-  message = new FormControl('',Validators.required);
+  name = new UntypedFormControl('',Validators.required);
+  email = new UntypedFormControl('',[Validators.required, Validators.email]);
+  message = new UntypedFormControl('',Validators.required);
 
 
   
